requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.43.0
httpx==0.27.2
requests==2.32.3
atproto==0.0.58
python-multipart==0.0.9
python-multipart==0.0.9
